% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_team.R
\name{filter_team}
\alias{filter_team}
\title{Pull all of the stats for any specific team}
\arguments{
\item{team_name}{Team name}

\item{season}{the year you want the information default is all seasons}
}
\description{
Get all of the information in the dataset for any team in any season
}
\examples{
filter_team("Utah St.")
filter_team("N.C. State", 2016)

}
