% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bracket_teams.R
\name{get_bracket_teams}
\alias{get_bracket_teams}
\title{Get NCAA Tournament Teams for a Given Year}
\arguments{
\item{yr}{An integer year to query.}

\item{data}{A data frame containing the dataset. Defaults to \code{cbb_data}.}
}
\value{
A sorted character vector of team names that made the NCAA bracket
in the given year.
}
\description{
Returns all teams that participated in the NCAA tournament for a given year.
This function is forward-compatible by assuming that any POSTSEASON label
other than "0", "N/A", or NA indicates a bracket team.
}
\examples{
get_bracket_teams(2023)
}
