% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_record.R
\name{get_team_record}
\alias{get_team_record}
\title{Extract a teams wins and losses in a given season}
\arguments{
\item{team_name}{String. The school's name}

\item{season_range}{Integer vector. A range of years (e.g., 2013:2024)}
}
\description{
Input a team name and range of years to get their totals wins and loses by
season
}
\examples{
get_team_record("Utah St.", 2014:2020)

}
