% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams_by_conf.R
\name{get_teams_by_conf}
\alias{get_teams_by_conf}
\title{Get Teams by Conference for a Given Year}
\usage{
get_teams_by_conf(year, data = cbb_data)
}
\arguments{
\item{year}{An integer year for which the user wants to retrieve
conference-team mappings.}

\item{data}{A data frame containing the dataset (must include 'year',
'team', and 'CONF' columns). Defaults to \code{DATSET}}
}
\value{
A named list where each element corresponds to a conference
and contains the vector of teams in that conference for the
given year.
}
\description{
This function returns a list of conferences and the teams that were
in each conference for a specified year.
}
\examples{
get_teams_by_conf(2018)
}
