% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power.R
\name{plot_power}
\alias{plot_power}
\title{See teams power rating over the years}
\arguments{
\item{team}{team name}
}
\description{
Input a team and you will see how their power ranking changed over the years
}
\examples{
plot_power("Utah St.")

}
