% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict_s16_teams}
\alias{predict_s16_teams}
\title{Predict the Sweet Sixteen}
\arguments{
\item{year}{A given year to describe the data}
}
\value{
a table of teams that we predict will mmake the sweet 16
}
\description{
This function takes the model and predicts the Sweet 16 for a
season's worth of data.
}
