% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator for getteamsbyconf.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate Inputs for get_teams_by_conf}
\usage{
validate_inputs(data, year)
}
\arguments{
\item{data}{The dataset to check.}

\item{year}{The year to validate.}
}
\value{
Stops with an error message if input is invalid; otherwise, returns TRUE invisibly.
}
\description{
Validate Inputs for get_teams_by_conf
}
