% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_team_record.R
\name{get_team_record}
\alias{get_team_record}
\title{Extract a teams wins and losses in a given season}
\usage{
get_team_record(team_name, season_range = NULL)
}
\arguments{
\item{team_name}{String. The school's name}

\item{season_range}{Integer vector. A range of years (e.g., 2013:2024)}
}
\description{
Extract a teams wins and losses in a given season
}
