% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator for getbracketteams.R
\name{validate_bracket_inputs}
\alias{validate_bracket_inputs}
\title{Validate Inputs for get_bracket_teams}
\usage{
validate_bracket_inputs(yr, data)
}
\arguments{
\item{data}{The dataset to check.}

\item{year}{The year to check.}
}
\value{
Stops with an error if invalid input is detected.
}
\description{
Validate Inputs for get_bracket_teams
}
